+++ lib/core/alice_dio_interceptor.dart
@@ -27,6 +27,7 @@ class AliceDioInterceptor extends InterceptorsWrapper {
     call.endpoint = path;
     call.server = uri.host;
     call.client = "Dio";
+    call.uri = options.uri.toString();
 
     if (uri.scheme == "https") {
       call.secure = true;