+++ lib/ui/alice_call_error_widger.dart
@@ -3,16 +3,26 @@ import 'package:flutter/material.dart';
 
 import 'alice_base_call_details_widget.dart';
 
-class AliceCallErrorWidget extends AliceBaseCallDetailsWidget {
+class AliceCallErrorWidget extends StatefulWidget {
   final AliceHttpCall call;
 
   AliceCallErrorWidget(this.call);
 
   @override
+  State<StatefulWidget> createState() {
+   return _AliceCallErrorWidgetState();
+  }
+}
+
+class _AliceCallErrorWidgetState
+    extends AliceBaseCallDetailsWidgetState<AliceCallErrorWidget> {
+  AliceHttpCall get _call => widget.call;
+
+  @override
   Widget build(BuildContext context) {
-    if (call.error != null) {
+    if (_call.error != null) {
       List<Widget> rows = List();
-      var error = call.error.error;
+      var error = _call.error.error;
       var errorText = "Error is empty";
       if (error != null) {
         errorText = error.toString();